Fixed bug in add.
git-svn-id: https://repo.eecs.berkeley.edu/svn/users/ksen/projects/corvette-instrumentor@6137 1c3fd406-9068-4342-8353-511fcdf43622

+++ src/InterpreterObserver.cpp
@@ -267,7 +267,7 @@ void InterpreterObserver::binop(IID iid, bool nuw, bool nsw, KVALUE* op1, KVALUE
   VALUE vresult;
   switch (op) {
     case ADD:
-      vresult.as_int = (int) v1 + (int) v2;
+      vresult.as_int = (uint64_t) v1 + (uint64_t) v2;
       break;
     case SUB:
       vresult.as_int = (int) v1 - (int) v2;
